Add support for plain text, only html and multipart emails.
The templates plugin already allows to return customized content
for the emails. Though, the email content is already sent as content
type html.
This improves it to allow sending only text, only html or multipart
emails. The templates plugin now returns a tuple of Option[Txt] and
Option[Html], so that clients can choose which parts they want to
provide. As the mailer plugin handles null or empty content gracefully
we can reduce None simply to null or an empty string (the latter is chosen
because the mailer plugin also passes an empty string if a part shall
be omitted).