Allow overriding Registration step redirect targets.
Registration actions like handleStartSignUp etc. redirected to login,
which is not always desirable. To allow users to change this
several properties are introduced (that can be omitted):
* onHandleStartSignUpGoTo
* onHandleSignUpGoTo
* onHandleStartResetPasswordGoTo
* onHandleResetPasswordGoTo

In securesocial/defaults.conf these are set to "/login", but this
could also be left out (if a property is not set it's set to
login either).

The properties based approach is chosen because there are already
an onLoginGoTo and onLogoutGoTo properties.