Support redirects to /login (triggered by SecuredAction) to use SSL
The ssl setting is used for the form action of the login page, and in
some other places. It might also be desirable to serve the login page
already using ssl (just to show users that they're already using ssl
and thus the login should also be secure).

Direct links to the login page are under the control of the application
and can be set accordingly. SecuredActions also redirect to the login
page, in the case of missing authentication, and in such a case it should
also be possible to switch/redirect to ssl.

This commit adds a new configuration property 'useSSLForLoginPage'
that is evaluated by SecureAction handlers (both scala and java). This
property can be seen as an extension of 'ssl', because it's an additional
requirement. Because of this, redirections to ssl are only done if
both ssl and useSSLForLoginPage are set to true.