On login/signUp redirect to the referer of the first signup/login request.
The referer is saved as original url when the login or signup page is
hit. If there's already an original url saved in the session, this is
not changed (e.g. the user might go to the login page, navigate to
signup and back again to login, then we want to keep the referer of the
first login request). To keep the request scheme/protocol, not the full
referer (including e.g. "http") is saved, but only the path and query
string. Otherwise, the user might be redirected from https back to
http and loose it's session (would not be logged in on http).

This changes the previous behaviour which used the onLoginGoTo property
to redirect the user to on login. This is now only used when no referer
was available.
If s.o. complains about this change one might change the behaviour so
that the referer is only used if no onLoginGoTo is configured.

Closes issue #157